crypto/[Hard] Blessed/requirements.txt
pycryptodome
py_ecc